+++ cds-plugin-vector-hana/lib/index.js
@@ -3,4 +3,5 @@ module.exports = {
   HanaVectorStore: require('./backends/hana'),
   SqliteVectorStore: require('./backends/sqlite'),
   RAG: require('./rag'),
+  activateCdsPlugin: require('./cdsPlugin').activate,
 };